# docs/apiSpec.yml
openapi: "3.1.1"
info:
  title: Degree Prerequisite Tracker
  description: "For all your degree planning needs!"
  version: "1.0"
  contact:
    email: degreeprerequisitetracker@gmail.com
...
paths:
...
  /users/history:
    get:
      tags:
        - Users
      summary: Get your course history
      description: Retrieves all of your completed courses.

      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    courseId:
                      type: integer
                      example: 1314
                    prefix:
                      type: string
                      example: CSC
                    number:
                      type: integer
                      example: 365

        "400":
          description: Invalid email or password.

      security:
        - bearerAuth: []

    put:
      tags:
        - Users
      summary: Adds courses to your course history
      description: Add the courseIds that you want to mark as completed.

      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: integer
              example: [1135, 1164, 1189]

      responses:
        "200":
          description: OK
        "400":
          description: Invalid email or password.

      security:
        - bearerAuth: []
  /users/history/{courseId}:
    delete:
      tags:
        - Users
      summary: Delete a course from your course history
      description: Deletes a specified course from your course history.
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
            example: 1314

      responses:
        "200":
          description: OK
        "400":
          description: Invalid email or password.

      security:
        - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...